chore: move constants.ts to lib directory
